新規登録時の処理をカスタムHook化
